separated private destination component from public destination component
