+++ frontend/src/Components/DestinationCard/DestinationCard.module.css
@@ -21,7 +21,7 @@
   z-index: 101;
   padding: 0.5rem;
   display: flex;
-  color: white;
+  color: black;
   flex-direction: column;
   justify-content: center;
   transition: 0.5s ease-in-out;
